+++ rana_qgis_plugin/rana_qgis_plugin.py
@@ -242,6 +242,9 @@ class RanaQgisPlugin:
             self.loader.loading_cancelled.connect(self.rana_browser.enable)
             self.rana_browser.download_file_selected.connect(self.loader.download_file)
             self.rana_browser.download_results_selected.connect(
+                self.rana_browser.disable
+            )
+            self.rana_browser.download_results_selected.connect(
                 self.loader.download_results
             )
             self.rana_browser.open_in_qgis_selected.connect(self.rana_browser.disable)
@@ -249,10 +252,8 @@ class RanaQgisPlugin:
             self.rana_browser.save_vector_styling_selected.connect(
                 self.rana_browser.disable
             )
+            self.loader.download_results_cancelled.connect(self.rana_browser.enable)
             self.rana_browser.download_file_selected.connect(self.rana_browser.disable)
-            self.rana_browser.download_results_selected.connect(
-                self.rana_browser.disable
-            )
             self.rana_browser.start_simulation_selected.connect(
                 self.rana_browser.disable
             )