+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 -------------------
 
 - Add revision view and new simulation button
+- Replace file interaction buttons with context menu
+- Download generated postprocessing rasters.
 
 
 1.1.20 (2025-11-18)